add `dev` flag to `exrm_reload`
It is added, for the case, that `ReleaseManager.Reload.run` can
reload configuration outside of release (for example in development
shell).

Aims to test the live reconfiguration of the application. With
combination of setting `watch` to true, it is possible to test
continius changing enviroments.

add in README the 5-liner code for testing the conform-generated
configuration in dev enviroment